new features and updates: - pipeline.tree.estimator.py removed and refactored to two files: estimator/props.py and estimator/main.py - Pipeline and BranchedPipeline structure refactored - execution report is now available in both pipeline and branched pipeline - execution report introduced: record actual quota usage - license removed - (fix) estimator: inaccurate execution report record resolved
