+++ .gitignore
@@ -19,3 +19,5 @@ explanation_draft.txt
 
 # documentations
 tests
+
+journey
