+++ yt_pipeline/pipeline/__init__.py
@@ -1,8 +1,10 @@
 """
 Chains multiple Foreman units into a multi-step workflow	
 """
-from .main import PipelineBlock, PipelineStacks, Pipeline, PipelineDeliverable
+from .main import Pipeline, PipelineDeliverable
+from .props import PipelineBlock, PipelineStacks
 from .constructor import PipelineStacksConstructor, PipelineBlocksConstructor, PipelineBlockConstructor
 from .estimator import PipelineEstimator
 
-from ..retriever.base import PipeSettings
+from .base import PipelineProduct, PipelineExecutionStage, PipelineExecutionReport, PipelineExecutionRecorder
+from ..retriever.base import PipeSettings, RetrieverSettings
