+++ yt_pipeline/regulator/__init__.py
@@ -3,4 +3,3 @@ Estimates quota cost for a given API call configuration
 """
 
 from .main import ChannelsRegulator, VideosRegulator, PlaylistItemsRegulator, PlaylistRegulator, SearchRegulator, CommentThreadsRegulator
-from .pipeline import PipelineRegulator
